+++ tests/modules/test_JMeterExecutor.py
@@ -798,7 +798,7 @@ class TestJMeterExecutor(BZTestCase):
         self.assertEqual(s_t('1m'), 60 * 1000.0)
         self.assertEqual(s_t('${VAR}'), '${VAR}')
 
-    def test_a1_json_body(self):
+    def test_json_body_app_str(self):
         obj = JMeterExecutor()
         obj.engine = EngineEmul()
         obj.execution.merge({
@@ -814,7 +814,7 @@ class TestJMeterExecutor(BZTestCase):
         selector += '>elementProp>stringProp[name="Argument.value"]'
         self.assertNotEqual(jmx.get(selector)[0].text.find('store_id'), -1)
 
-    def test_a2_json_body(self):
+    def test_json_body_app_dic(self):
         obj = JMeterExecutor()
         obj.engine = EngineEmul()
         obj.execution.merge({
@@ -832,7 +832,7 @@ class TestJMeterExecutor(BZTestCase):
         selector += '>elementProp>stringProp[name="Argument.value"]'
         self.assertNotEqual(jmx.get(selector)[0].text.find('store_id'), -1)
 
-    def test_a3_json_body(self):
+    def test_json_body_no_app(self):
         obj = JMeterExecutor()
         obj.engine = EngineEmul()
         obj.execution.merge({
@@ -849,3 +849,18 @@ class TestJMeterExecutor(BZTestCase):
         selector = 'elementProp[name="HTTPsampler.Arguments"]>collectionProp'
         selector += '>elementProp>stringProp[name="Argument.value"]'
         self.assertEqual(jmx.get(selector)[0].text.find('store_id'), -1)
+
+    def test_a1_jtl_verbose(self):
+        obj = JMeterExecutor()
+        obj.engine = EngineEmul()
+        obj.execution.merge({
+            "write-xml-jtl": "full",
+            "scenario": {
+                "requests": [{
+                    "url": "http://blazedemo.com",
+                    "headers": {"Content-Type": "application/exi"},
+                    "body": {
+                        "store_id": "${store_id}",
+                        "display_name": "${display_name}"
+                    }}]}})
+        obj.prepare()